fix(core,urql): useDebugState and some envvar handling
