+++ src/urql/clientAtom.ts
@@ -2,7 +2,12 @@ import { createClient } from '@urql/core'
 import { atom } from 'jotai'
 
 const DEFAULT_URL =
-  (typeof process === 'object' && process.env.JOTAI_URQL_DEFAULT_URL) ||
-  '/graphql'
+  (() => {
+    try {
+      return process.env.JOTAI_URQL_DEFAULT_URL
+    } catch {
+      return undefined
+    }
+  })() || '/graphql'
 
 export const clientAtom = atom(createClient({ url: DEFAULT_URL }))
