+++ tests/async.test.tsx
@@ -6,6 +6,8 @@ import { getTestProvider, itSkipIfVersionedWrite } from './testUtils'
 
 const Provider = getTestProvider()
 
+jest.mock('../src/core/useDebugState.ts')
+
 const useCommitCount = () => {
   const commitCountRef = useRef(1)
   useEffect(() => {
