+++ tests/utils/atomFamily.test.tsx
@@ -8,6 +8,8 @@ import { getTestProvider } from '../testUtils'
 
 const Provider = getTestProvider()
 
+jest.mock('../../src/core/useDebugState.ts')
+
 // FIXME this is a hacky workaround temporarily
 const IS_REACT18 = !!(ReactDOM as any).createRoot
